+++ src/main/java/com/company/math/vector/Vector4.java
@@ -3,12 +3,12 @@ package com.company.math.vector;
 import java.util.Objects;
 
 public class Vector4 implements Vector<Vector4> {
-    public double x;
-    public double y;
-    public double z;
-    public double k;
+    public float x;
+    public float y;
+    public float z;
+    public float k;
 
-    public Vector4(double x, double y, double z, double k) {
+    public Vector4(float x, float y, float z, float k) {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -16,8 +16,8 @@ public class Vector4 implements Vector<Vector4> {
     }
 
     @Override
-    public double length(){
-        return Math.sqrt(x * x + y * y + z * z + k * k);
+    public float length(){
+        return (float) Math.sqrt(x * x + y * y + z * z + k * k);
     }
 
     @Override
@@ -31,14 +31,14 @@ public class Vector4 implements Vector<Vector4> {
     }
     
 
-    public void multiplyingAVectorByAScalar(double n){
+    public void multiplyingAVectorByAScalar(float n){
         x *= n;
         y *= n;
         z *= n;
         k *= n;
     }
 
-     public double scalarProduct(Vector4 v){
+     public float scalarProduct(Vector4 v){
         return (x * v.x + y * v.y + z * v.z + k * v.k)/length()/v.length();
     }
 
@@ -47,7 +47,7 @@ public class Vector4 implements Vector<Vector4> {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Vector4 vector4 = (Vector4) o;
-        return Double.compare(vector4.x, x) == 0 && Double.compare(vector4.y, y) == 0 && Double.compare(vector4.z, z) == 0 && Double.compare(vector4.k, k) == 0;
+        return Float.compare(vector4.x, x) == 0 && Float.compare(vector4.y, y) == 0 && Float.compare(vector4.z, z) == 0 && Float.compare(vector4.k, k) == 0;
     }
 
     @Override
